Attempted to implement closed-deceased, not exactly working, but my commit doesn't break our code.
What confused me for a long time is that in the issues it was suggested that we modify booking status from room.php and it was suggested in class that it was via dropdown box, but I scoured the codebase for a dropdown indicating status.  I found what I think we were talking about, and it referred to the other type of status (dirty room, etc.) and am unsure whether we want a new box from scratch/what to call it, or if we want to have checking out done without a dropdown box and perhaps something like a checkmark box.

The other issue I'm having is that data.php is not displaying my added closed bookings column in the first table.  I cannot figure out why this is, and have not implemented the remainder of the tables though it would hypothetically be just a lot of copy/paste.